rfid_dashboard/requirements.txt
Flask
pyserial 